+++ src/main/java/kr/hhplus/be/server/reservation/domain/ReservationStatus.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.reservation;
+package kr.hhplus.be.server.reservation.domain;
 
 public enum ReservationStatus {
